api/requirements.txt
fastapi
uvicorn
numpy<2
pillow
tensorflow<2.11
python-multipart